refactor: move OAuth routes to blueprint in oauth.py
- Created oauth_bp blueprint with url_prefix '/api/v1/auth/oauth'
- Moved all OAuth routes from __init__.py to oauth.py blueprint
- Registered blueprint in create_app() function
- Cleaned up unused imports in __init__.py
- Maintained all existing OAuth functionality and URL structure
